fix(matrix): refactoring transport http handler
